+++ PoutineStyle/StatePlayGame.cpp
@@ -4,7 +4,7 @@ StatePlayGame::StatePlayGame(GameManager* game, Settings settings, std::string m
 {
     // Settings
     this->settings = settings;
-    this->speedFactor = this->settings.getSensibility();
+    this->speedFactor = float(this->settings.getSensibility());
 
     if (this->settings.getDifficulty() == 2)
     {
@@ -100,8 +100,8 @@ sf::Vector2f StatePlayGame::rotateVectorMatrix(sf::Vector2f v, double a) {
     // [cosa  -sina]
     // [sina   cosa]
     sf::Vector2f resVec = sf::Vector2f();
-    resVec.x = v.x * cos(a) - v.y * sin(a);
-    resVec.y = v.x * sin(a) + v.y * cos(a);
+    resVec.x = float(v.x * cos(a) - v.y * sin(a));
+    resVec.y = float(v.x * sin(a) + v.y * cos(a));
 
     // Need a unit vector --> Divide vector's components by the length of the vector
     /*double vecLen = sqrt(pow(resVec.x, 2) + pow(resVec.y, 2));
@@ -289,10 +289,10 @@ void StatePlayGame::updatePlayerPosition(sf::Vector2f newPos)
 {
     int newPosMapY, newPosMapX, oldPosMapX, oldPosMapY;
 
-    oldPosMapY = floor(player->position.x);
-    newPosMapY = newPos.x;
-    oldPosMapX = floor(player->position.y);
-    newPosMapX = newPos.y;
+    oldPosMapY = int(floor(player->position.x));
+    newPosMapY = int(newPos.x);
+    oldPosMapX = int(floor(player->position.y));
+    newPosMapX = int(newPos.y);
 
     // Check X axis
     if (!Entity::isWall(map[newPosMapX][oldPosMapY]))
@@ -343,20 +343,20 @@ void StatePlayGame::endGameManagment()
 void StatePlayGame::drawMap2D()
 {
     sf::RectangleShape block;
-    block.setSize(sf::Vector2f(blockWidth, blockHeight));
+    block.setSize(sf::Vector2f(float(blockWidth), float(blockHeight)));
     block.setOutlineThickness(1);
     block.setOutlineColor(sf::Color::Black);
 
     sf::CircleShape player_circle;
     player_circle.setRadius(5);
-    player_circle.setPosition(sf::Vector2f(player->position.x * (blockWidth) - player_circle.getRadius(), player->position.y * (blockHeight - 1) - player_circle.getRadius() / 2.0));
+    player_circle.setPosition(sf::Vector2f(player->position.x * (blockWidth) - player_circle.getRadius(), player->position.y * (blockHeight - 1.f) - player_circle.getRadius() / 2.f));
     player_circle.setFillColor(sf::Color::Green);
 
     for (int i = 0; i < mapSize; i++)
     {
         for (int j = 0; j < mapSize; j++)
         {
-            block.setPosition(j * blockWidth, i * blockHeight);
+            block.setPosition(float(j * blockWidth), float(i * blockHeight));
 
             if (map[i][j] == '1') // needs a fastest way to test if there's more cases. [ O(n^2) ]
             {
@@ -396,18 +396,18 @@ void StatePlayGame::drawMiniMap()
 
     sf::RectangleShape rectBackground = sf::RectangleShape();
     rectBackground.setFillColor(sf::Color::White);
-    rectBackground.setPosition(blockWUnit - 2, 2 * blockHUnit - 2); // problème ici
-    rectBackground.setSize(sf::Vector2f(7* blockWUnit + 4, 7* blockHUnit + 4));
+    rectBackground.setPosition(blockWUnit - 2.f, 2 * blockHUnit - 2.f);
+    rectBackground.setSize(sf::Vector2f(7.f * blockWUnit + 4.f, 7.f * blockHUnit + 4.f));
     gameManager->getRenderWindow()->draw(rectBackground);
 
     for (int x = -3; x <= 3; x++) {
         for (int y = -3; y <= 3; y++) {
-            int positionX = floor(player->position.x) + x;
-            int positionY = floor(player->position.y) + y;
+            int positionX = int(floor(player->position.x) + x);
+            int positionY = int(floor(player->position.y) + y);
             
             sf::RectangleShape cell = sf::RectangleShape();
-            cell.setPosition(sf::Vector2f((3 - x) * blockWUnit + blockWUnit, (4 - y) * blockHUnit + blockHUnit));
-            cell.setSize(sf::Vector2f(blockWUnit, blockHUnit));
+            cell.setPosition(sf::Vector2f((3.f - x) * blockWUnit + blockWUnit, (4.f - y) * blockHUnit + blockHUnit));
+            cell.setSize(sf::Vector2f(float(blockWUnit), float(blockHUnit)));
             
             if (positionX >= 0 && positionX < mapSize && positionY >= 0 && positionY < mapSize) {
                 char charCell = map[positionY][positionX];
@@ -436,8 +436,8 @@ void StatePlayGame::drawMiniMap()
     // Draw player
     sf::CircleShape circle = sf::CircleShape();
     circle.setFillColor(sf::Color::Green);
-    circle.setPosition(3 * blockWUnit + blockWUnit, 4 * blockHUnit + blockHUnit);
-    circle.setRadius(blockWUnit / 2);
+    circle.setPosition(3.f * blockWUnit + blockWUnit, 4.f * blockHUnit + blockHUnit);
+    circle.setRadius(blockWUnit / 2.f);
     gameManager->getRenderWindow()->draw(circle);
 
     // Draw player direction vector
@@ -447,9 +447,9 @@ void StatePlayGame::drawMiniMap()
     rect.setFillColor(sf::Color::Green);
 
     // Update rotation
-    double delta = player->direction.x / player->direction.y;
-    double angleInRad = atan(delta);
-    double angleInDegrees = angleInRad * 180.0 / 3.1415 + 180.0; // Use a constant
+    float delta = player->direction.x / player->direction.y;
+    float angleInRad = atan(delta);
+    float angleInDegrees = angleInRad * 180.f / float(3.1415) + 180.f; // Use a constant
     if (player->direction.y < 0) angleInDegrees += 180.0;
     rect.rotate(-angleInDegrees);
     gameManager->getRenderWindow()->draw(rect);
@@ -482,7 +482,7 @@ void StatePlayGame::castRay(int x, int w, int depth)
     // Vector representing the direction of the actual ray
     double cameraX = double(2.f * double(x)) / double(w) - 1.0;
 
-    sf::Vector2f rayDir = player->direction + sf::Vector2f(player->planeVec.x * cameraX, player->planeVec.y * cameraX);
+    sf::Vector2f rayDir = player->direction + sf::Vector2f(player->planeVec.x * float(cameraX), player->planeVec.y * float(cameraX));
     double rayDirLen = std::sqrt(pow(rayDir.x, 2) + pow(rayDir.y, 2));
 
     // Fisheye effect
@@ -591,7 +591,7 @@ void StatePlayGame::castRay(int x, int w, int depth)
         sf::Vector2i texture_coords = currentWall->getCurrentTextureCoordinates() * textureSize;
 
         // Calculate where the wall was hit
-        float wallX;
+        double wallX;
         if (isWallHitHorizontal) wallX = (player->position.y) + perpWallDist * rayDir.y;
         else wallX = (player->position.x) + perpWallDist * rayDir.x;
         wallX -= floor(wallX);
@@ -767,8 +767,8 @@ void StatePlayGame::renderingEntities(double dt) {
         if (typeid(*entity).name() == typeid(Bullet).name()) {
             Bullet* bullet = static_cast<Bullet*>(entity);
             // Check if the bullet hit something
-            int nextX = floor((float)bullet->mapPos.x + 0.5 + bullet->getVelocity().x * dt * 10.f);
-            int nextY = floor((float)bullet->mapPos.y + 0.5 + bullet->getVelocity().y * dt * 10.f);
+            int nextX = int(floor((float)bullet->mapPos.x + 0.5 + bullet->getVelocity().x * dt * 10.f));
+            int nextY = int(floor((float)bullet->mapPos.y + 0.5 + bullet->getVelocity().y * dt * 10.f));
 
             if (nextX < 0 || nextY < 0 || nextX > mapSize - 1 || nextY > mapSize - 1) {
                 bullet->isTravelling = false;
@@ -875,11 +875,11 @@ void StatePlayGame::renderingEntities(double dt) {
         if (typeid(*entity).name() == typeid(Guard).name() || typeid(*entity).name() == typeid(General).name()) {
             Ennemy* ennemy = static_cast<Ennemy*>(entity);
             // Calculate the direction of the bullet (aiming the player)
-            sf::Vector2f bulletDir = sf::Vector2f(player->position.x - 0.5, player->position.y - 0.5) - ennemy->mapPos;
+            sf::Vector2f bulletDir = sf::Vector2f(player->position.x - 0.5f, player->position.y - 0.5f) - ennemy->mapPos;
             // Get the norm of the direction vector
             double norm = sqrt(pow(bulletDir.x, 2) + pow(bulletDir.y, 2));
             // Get the unit vector
-            sf::Vector2f bulletDirUnit = sf::Vector2f(bulletDir.x / norm, bulletDir.y / norm);
+            sf::Vector2f bulletDirUnit = sf::Vector2f(bulletDir.x / float(norm), bulletDir.y / float(norm));
             std::stack<Bullet*> bullets = ennemy->shoot(bulletDirUnit, this->player->position, this->map);
             while (!bullets.empty()) {
                 Bullet* bullet = bullets.top();
@@ -888,14 +888,14 @@ void StatePlayGame::renderingEntities(double dt) {
             }
         }
         entity->draw(*gameManager->getRenderWindow(), player->position, player->direction, player->planeVec, ZBuffer, gameManager->getWindowWidth(), gameManager->getWindowHeight());
-        entity->update(dt); // Update the animation
+        entity->update(float(dt)); // Update the animation
     }
 
 #pragma endregion
 
 #pragma region Rendering player sprites
     player->draw(*gameManager->getRenderWindow());
-    player->update(dt);
+    player->update(float(dt));
     showCursor();
 #pragma endregion
 }
@@ -920,7 +920,7 @@ void StatePlayGame::parseMap2D()
             map[indexX][indexY] = *it;
             if (map[indexX][indexY] == 'T') // Player spawn position
             {
-                player->position = sf::Vector2f(indexY + 0.5, indexX + 0.5);
+                player->position = sf::Vector2f(indexY + 0.5f, indexX + 0.5f);
             }
             else if (map[indexX][indexY] == 'E') { // Ennemy
                 int proba[] = {25, 25, 50};
@@ -1106,12 +1106,12 @@ void StatePlayGame::setHud()
     sf::Color lineCol = sf::Color::Red;
     sf::Color hudBack = sf::Color::White;
 
-    hudUp->setSize(sf::Vector2f(gameManager->getWindowWidth(), gameManager->getWindowHeight() / 25));
-    hudDownL->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10));
-    hudDownML->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10));
-    hudDownM->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10));
-    hudDownMR->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10));
-    hudDownR->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10));
+    hudUp->setSize(sf::Vector2f(float(gameManager->getWindowWidth()), gameManager->getWindowHeight() / 25.f));
+    hudDownL->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10.f));
+    hudDownML->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10.f));
+    hudDownM->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10.f));
+    hudDownMR->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10.f));
+    hudDownR->setSize(sf::Vector2f(hudDownWidth, gameManager->getWindowHeight() / 10.f));
 
     hudUp->setFillColor(hudBack);
     hudUp->setOutlineColor(lineCol);
@@ -1121,27 +1121,27 @@ void StatePlayGame::setHud()
     hudDownL->setFillColor(hudBack);
     hudDownL->setOutlineColor(lineCol);
     hudDownL->setOutlineThickness(3);
-    hudDownL->setPosition(3, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10 - 3);
+    hudDownL->setPosition(3.f, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10.f - 3.f);
 
     hudDownML->setFillColor(hudBack);
     hudDownML->setOutlineColor(lineCol);
     hudDownML->setOutlineThickness(3);
-    hudDownML->setPosition(hudDownWidth + 3, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10 - 3);
+    hudDownML->setPosition(hudDownWidth + 3.f, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10.f - 3.f);
 
     hudDownM->setFillColor(hudBack);
     hudDownM->setOutlineColor(lineCol);
     hudDownM->setOutlineThickness(3);
-    hudDownM->setPosition(hudDownWidth * 2 + 6, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10 - 3);
+    hudDownM->setPosition(hudDownWidth * 2.f + 6.f, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10.f - 3.f);
 
     hudDownMR->setFillColor(hudBack);
     hudDownMR->setOutlineColor(lineCol);
     hudDownMR->setOutlineThickness(3);
-    hudDownMR->setPosition(hudDownWidth * 3 + 9, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10 - 3);
+    hudDownMR->setPosition(hudDownWidth * 3.f + 9.f, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10.f - 3.f);
 
     hudDownR->setFillColor(hudBack);
     hudDownR->setOutlineColor(lineCol);
     hudDownR->setOutlineThickness(3);
-    hudDownR->setPosition(hudDownWidth * 4 + 12, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10 - 3);
+    hudDownR->setPosition(hudDownWidth * 4.f + 12.f, gameManager->getWindowHeight() - gameManager->getWindowHeight() / 10.f - 3.f);
 }
 
 void StatePlayGame::displayHud()
@@ -1155,8 +1155,8 @@ void StatePlayGame::displayHud()
     gameManager->getRenderWindow()->draw(*hudDownR);
 
     float widthPos = 1.f;
-    float hudDownWidth = widthPos * (gameManager->getWindowWidth() / 5) - 3;
-    float heightPos = gameManager->getWindowHeight() / gameManager->getWindowHeight();
+    float hudDownWidth = widthPos * (gameManager->getWindowWidth() / 5.f) - 3.f;
+    float heightPos = float(gameManager->getWindowHeight()) / gameManager->getWindowHeight();
 
     sf::Font font = gameManager->getFont();
     sf::Color fontCol = sf::Color::Red;
@@ -1230,9 +1230,9 @@ void StatePlayGame::showCursor()
     sf::Vector2u cursorSize = imgAimCursor.getSize();
     int cursorWidth = cursorSize.x, cursorHeigth = cursorSize.y;
 
-    sf::Vector2f centerWindowPos = sf::Vector2f(this->gameManager->getWindowWidth() / 2, this->gameManager->getWindowHeight() / 2);
+    sf::Vector2f centerWindowPos = sf::Vector2f(this->gameManager->getWindowWidth() / 2.f, this->gameManager->getWindowHeight() / 2.f);
 
-    aimCursor.setPosition(centerWindowPos.x - (cursorWidth / 4.5), centerWindowPos.y - (cursorHeigth / 4));
+    aimCursor.setPosition(centerWindowPos.x - (cursorWidth / 4.5f), centerWindowPos.y - (cursorHeigth / 4.f));
 
     this->gameManager->getRenderWindow()->draw(aimCursor);
 }
@@ -1280,8 +1280,8 @@ int StatePlayGame::lootManagment(int* proba, int length)
 /// </summary>
 /// <returns>Return * Entity or nullptr</returns>
 Entity* StatePlayGame::getInteractedEntity() {
-    int playerDirectionPosX = floor(player->position.x + 1.0 * player->direction.x);
-    int playerDirectionPosY = floor(player->position.y + 1.0 * player->direction.y);
+    int playerDirectionPosX = int(floor(player->position.x + 1.0 * player->direction.x));
+    int playerDirectionPosY = int(floor(player->position.y + 1.0 * player->direction.y));
     if (playerDirectionPosX > 0 && playerDirectionPosX < mapSize && playerDirectionPosY > 0 && playerDirectionPosY < mapSize) {
         if (this->entityMap[playerDirectionPosX][playerDirectionPosY] != nullptr) return entityMap[playerDirectionPosX][playerDirectionPosY];
     }
